requirements.txt
flask==2.0.1
tensorflow-cpu==2.9.1
numpy==1.23.5
pandas==1.5.3
matplotlib==3.5.3
seaborn==0.12.2
scikit-learn==1.0.2
gunicorn==20.1.0 